Pylint local transform plugin proxy
On dynmaic classes, and modules pylint has hard time to infer whether
dynamically added attributes are valid or not. Through the --load-plugin
command line parameter and the load-plugin configuraiton file parameter
additional modules providing the appropriate knowledge can be loaded and
activated.

The disadvantage is that on complex projects all plugins have to be
specified either every time on commandline or permanently in
configuration file. Which is not allways desired or tedious.

The plugin proxy module allows to store additional plugin modules along
with the module it should provide transformatoin and type infrerences
for. In case the module defining a monitored astroid node type is
encountered for the first time the corresponding <modulename>.ast.py
plugin module is loaded if present on the path the <module>.py file was
loaded from.